categories - set up subscribe button container
